requirements.txt
PyPDF2==3.0.1
google-generativeai==0.3.1
python-dotenv==1.0.0
openai==1.10.0